--- westcoast/src/main/java/frc/subsystems/drive/motorfactory/MotorFactory.kt
@@ -1,7 +1,0 @@
-package frc.subsystems.drive.motorfactory
-
-import edu.wpi.first.wpilibj.SpeedController
-
-interface MotorFactory {
-  fun create(channelMain: Int, channel1: Int, channel2: Int): SpeedController
-}
