+++ westcoast/src/main/kotlin/frc/commands/auto/CenterAutonomous.kt
@@ -4,13 +4,10 @@ import edu.wpi.first.wpilibj.Encoder
 import edu.wpi.first.wpilibj.drive.DifferentialDrive
 import edu.wpi.first.wpilibj2.command.SequentialCommandGroup
 import edu.wpi.first.wpilibj2.command.WaitCommand
-import frc.commands.auto.demo.Square
-import edu.wpi.first.wpilibj2.command.CommandBase
-import edu.wpi.first.wpilibj.controller.ProfiledPIDController
 import frc.robot.maps.RobotMap
 import frc.subsystems.BallHandlingSubsystem
-import frc.subsystems.vision.VisionRunner
-import frc.subsystems.vision.targeting.AlignToTarget
+import frc.subsystems.VisionRunner
+import frc.subsystems.targeting.AlignToTarget
 
 class CenterAutonomous(
     map: RobotMap,
