+++ westcoast/src/main/kotlin/frc/commands/auto/LoadingBayAutonomous.kt
@@ -6,19 +6,22 @@ import edu.wpi.first.wpilibj2.command.SequentialCommandGroup
 import edu.wpi.first.wpilibj2.command.WaitCommand
 import frc.robot.maps.RobotMap
 import frc.subsystems.BallHandlingSubsystem
-import frc.subsystems.vision.VisionRunner
-import frc.subsystems.vision.targeting.AlignToTarget
+import frc.subsystems.VisionRunner
+import frc.subsystems.targeting.AlignToTarget
 
-class TargetAutonomous(
+class LoadingBayAutonomous(
     map: RobotMap,
     drive: DifferentialDrive,
     leftEncoder: Encoder,
     rightEncoder: Encoder,
     distance: Double,
+    angle: Double,
     vision: VisionRunner?,
     ballHandler: BallHandlingSubsystem?
 ) : SequentialCommandGroup(
     LeaveLine(map, drive, leftEncoder, rightEncoder, distance),
+    WaitCommand(1.0),
+    PerfectTurn(map, drive, leftEncoder, rightEncoder, angle),
     WaitCommand(1.0),
     AlignToTarget(map, drive, vision!!)
 )
