+++ westcoast/src/main/kotlin/frc/math/ShooterPower.kt
