+++ westcoast/src/main/kotlin/frc/robot/ShinDestroyer.kt
@@ -11,28 +11,32 @@ import edu.wpi.first.wpilibj.smartdashboard.SendableChooser
 import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard
 import edu.wpi.first.wpilibj2.command.Command
 import edu.wpi.first.wpilibj2.command.CommandScheduler
+import edu.wpi.first.wpilibj2.command.SequentialCommandGroup
 import edu.wpi.first.wpilibj2.command.button.JoystickButton
 import frc.commands.auto.CenterAutonomous
 import frc.commands.auto.LoadingBayAutonomous
 import frc.commands.auto.TargetAutonomous
+import frc.commands.auto.demo.FlexAuto
 import frc.commands.motors.SetMotor
-import frc.robot.maps.ClinkyMap
 import frc.robot.maps.RobotMap
+import frc.robot.maps.ShinDestroyerMap
 import frc.subsystems.BallHandlingSubsystem
 import frc.subsystems.drive.GyroDrive
 import frc.subsystems.drive.motorfactory.MotorFactory
 import frc.subsystems.drive.motorfactory.MotorFactoryHybrid
-import frc.subsystems.vision.RioVisionRunner
-import frc.subsystems.vision.VisionRunner
+import frc.subsystems.LimelightRunner
+import frc.subsystems.VisionRunner
+import frc.subsystems.targeting.AlignToTarget
+import frc.subsystems.targeting.FindTarget
 
-class Clinky : TimedRobot() {
+class ShinDestroyer : TimedRobot() {
   // Robot Map
-  private val map: RobotMap = ClinkyMap()
+  private val map: RobotMap = ShinDestroyerMap()
 
   // Drivetrain
   var hybridFactory: MotorFactory = MotorFactoryHybrid()
-  private val leftMain = hybridFactory.create(2, 1, 3)
-  private val rightMain = hybridFactory.create(5, 4, 6)
+  private val leftMain = hybridFactory.create(14, 13, 15)
+  private val rightMain = hybridFactory.create(2, 1, 3)
   private val drive = DifferentialDrive(leftMain, rightMain)
 
   // Subsystem Motors
@@ -44,11 +48,11 @@ class Clinky : TimedRobot() {
 
   // Sensors
   private val ahrs = AHRS(SPI.Port.kMXP)
-  private val leftEncoder = Encoder(0, 1)
-  private val rightEncoder = Encoder(3, 2)
+  private val leftEncoder = Encoder(9, 8)
+  private val rightEncoder = Encoder(6, 7)
 
   // Vision
-  private val vision: VisionRunner = RioVisionRunner()
+  private val vision: VisionRunner = LimelightRunner()
 
   // Inputs
   private val joystick = Joystick(0)
@@ -62,26 +66,31 @@ class Clinky : TimedRobot() {
 
   // Autonomous Commands
   private val autonomousChooser = SendableChooser<Command>()
-  private var autonomous: Command? = null
-  private val autonTarget = TargetAutonomous(map, drive, leftEncoder, rightEncoder, 3.0, vision, ballHandler)
-  private val autonCenter = CenterAutonomous(map, drive, leftEncoder, rightEncoder, 3.0, 30.0, vision, ballHandler)
-  private val autonBay = LoadingBayAutonomous(map, drive, leftEncoder, rightEncoder, 3.0, -40.0, vision, ballHandler)
+  private var autonomous: Command = TargetAutonomous(map, drive, leftEncoder, rightEncoder, -3.0, vision, ballHandler)
+  private val autonTarget = TargetAutonomous(map, drive, leftEncoder, rightEncoder, -3.0, vision, ballHandler)
+  private val autonCenter = CenterAutonomous(map, drive, leftEncoder, rightEncoder, -3.0, -30.0, vision, ballHandler)
+  private val autonBay = LoadingBayAutonomous(map, drive, leftEncoder, rightEncoder, -3.0, -40.0, vision, ballHandler)
+  private val spinnyCommand: Command = SequentialCommandGroup(FindTarget(drive, vision), AlignToTarget(map, drive, vision))
+  private val phlexAuto = FlexAuto(map, drive, leftEncoder, rightEncoder, vision, ballHandler)
   override fun robotInit() {
     drive.isSafetyEnabled = false
+    leftEncoder.distancePerPulse = 0.390625
+    rightEncoder.distancePerPulse = 0.390625
     autonomousChooser.setDefaultOption("Target", autonTarget)
     autonomousChooser.setDefaultOption("Center", autonCenter)
     autonomousChooser.setDefaultOption("LoadingBay", autonBay)
+    autonomousChooser.setDefaultOption("SpinFast", spinnyCommand)
+    autonomousChooser.setDefaultOption("Flex on Mr. Felipe", phlexAuto)
     SmartDashboard.putData("Autonomous Mode", autonomousChooser)
   }
 
   override fun robotPeriodic() {}
   override fun autonomousInit() {
-    CommandScheduler.getInstance().cancelAll()
     autonomous = autonomousChooser.selected
     ahrs.reset()
     leftEncoder.reset()
     rightEncoder.reset()
-    autonomous?.schedule()
+    autonomous.schedule()
   }
 
   override fun autonomousPeriodic() {
@@ -89,13 +98,13 @@ class Clinky : TimedRobot() {
   }
 
   override fun teleopInit() {
-    autonomous!!.cancel()
+    autonomous.cancel()
     teleop.cancel()
     ahrs.reset()
     leftEncoder.reset()
     rightEncoder.reset()
     teleop.schedule()
-    topTrigger.whileHeld(SetMotor(intake, 0.4))
+    topTrigger.whenPressed(SetMotor(intake, 1.0))
   }
 
   override fun teleopPeriodic() {
