+++ westcoast/src/main/kotlin/frc/robot/maps/ShinDestroyerMap.kt
@@ -1,53 +1,12 @@
 package frc.robot.maps
 
-import frc.math.Maths
-import frc.robot.maps.RobotMap
 import edu.wpi.first.wpilibj.controller.ProfiledPIDController
 import edu.wpi.first.wpilibj.trajectory.TrapezoidProfile
 import edu.wpi.first.wpilibj.DigitalInput
 import com.ctre.phoenix.motorcontrol.can.WPI_TalonSRX
 import edu.wpi.first.wpilibj.SpeedController
 import com.ctre.phoenix.motorcontrol.can.WPI_VictorSPX
-import edu.wpi.first.wpilibj.VictorSP
-import kotlin.jvm.JvmStatic
-import edu.wpi.first.wpilibj.RobotBase
-import frc.robot.Dinky
-import java.util.function.Supplier
-import edu.wpi.first.wpilibj.TimedRobot
-import frc.robot.maps.DinkyMap
-import frc.subsystems.drive.motorfactory.MotorFactory
-import frc.subsystems.drive.motorfactory.MotorFactoryHybrid
-import edu.wpi.first.wpilibj.drive.DifferentialDrive
-import frc.subsystems.BallHandlingSubsystem
-import frc.subsystems.vision.RioVisionRunner
-import edu.wpi.first.wpilibj.Joystick
-import edu.wpi.first.wpilibj2.command.button.JoystickButton
-import frc.subsystems.drive.TeleopDrive
-import edu.wpi.first.wpilibj2.command.CommandScheduler
-import frc.commands.ballhandling.IntakeAndIndex
-import frc.commands.ballhandling.IntakeOut
-import frc.commands.ballhandling.EverythingOut
-import frc.commands.ballhandling.Shooting
-import frc.robot.maps.JankyMap
-import edu.wpi.first.wpilibj.Spark
-import com.analog.adis16448.frc.ADIS16448_IMU
-import edu.wpi.first.networktables.NetworkTableInstance
-import edu.wpi.first.networktables.NetworkTableEntry
-import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard
-import frc.robot.maps.ClinkyMap
-import com.kauailabs.navx.frc.AHRS
-import edu.wpi.first.wpilibj.SPI
 import edu.wpi.first.wpilibj.controller.PIDController
-import frc.subsystems.drive.GyroDrive
-import edu.wpi.first.wpilibj.smartdashboard.SendableChooser
-import frc.commands.auto.TargetAutonomous
-import frc.commands.auto.CenterAutonomous
-import frc.commands.auto.LoadingBayAutonomous
-import frc.commands.motors.SetMotor
-import frc.robot.maps.ShinDestroyerMap
-import frc.subsystems.vision.LimelightRunner
-import edu.wpi.first.wpilibj2.command.SequentialCommandGroup
-import frc.commands.auto.demo.FlexAuto
 
 class ShinDestroyerMap : RobotMap {
   override fun makeLLPIDs(): ProfiledPIDController {
