+++ westcoast/src/main/kotlin/frc/subsystems/drive/motorfactory/MotorFactory.kt
@@ -1,0 +1,7 @@
+package frc.subsystems.drive.motorfactory
+
+import edu.wpi.first.wpilibj.SpeedController
+
+interface MotorFactory {
+  fun create(channelMain: Int, channel1: Int, channel2: Int): SpeedController
+}
