+++ westcoast/src/main/kotlin/frc/subsystems/targeting/StayOnTarget.kt
@@ -1,34 +1,30 @@
-package frc.subsystems.vision.targeting
+package frc.subsystems.targeting
 
 import edu.wpi.first.wpilibj.controller.PIDController
 import edu.wpi.first.wpilibj.drive.DifferentialDrive
 import edu.wpi.first.wpilibj2.command.CommandBase
 import frc.math.Maths.clamp
 import frc.robot.maps.RobotMap
-import frc.subsystems.vision.VisionRunner
+import frc.subsystems.VisionRunner
 
-class ConstantlyAlignToTarget(
+class StayOnTarget(
     private val map: RobotMap,
     private val drive: DifferentialDrive,
     private val vision: VisionRunner
 ) : CommandBase() {
-  private var turn: PIDController? = null
-  private var forward: PIDController? = null
+  private var pid: PIDController? = null
   override fun initialize() {
-    turn = map.makeTurnCAlignPIDs()
-    forward = map.makeForwardCAlignPIDs()
+    pid = map.makeStayOnTargetPIDs()
   }
 
   override fun execute() {
     val value = vision.alignX
-    val distance = vision.distance
-    val turnpower = clamp(turn!!.calculate(value, 0.0), 0.4)
-    val forwardpower = clamp(forward!!.calculate(distance, 5.0), 0.4)
-    drive.arcadeDrive(-forwardpower, turnpower, false)
+    val turn = clamp(pid!!.calculate(value, 0.0), 0.4)
+    drive.arcadeDrive(0.0, turn, false)
   }
 
   override fun end(interrupted: Boolean) {
-    turn!!.reset()
+    pid!!.reset()
     drive.stopMotor()
   }
